+++ huffman.h
@@ -18,9 +18,9 @@
 #define INSERT_ITEM    1                    
 #define SWITCH_ITEMS   2                    // Switch the item1 and item2
 
-#define PTR_NOT(ptr)  (THTreeItem *)(~(unsigned long)(ptr))
+#define PTR_NOT(ptr)  (THTreeItem *)(~(uintptr_t)(ptr))
 #define PTR_PTR(ptr)  ((THTreeItem *)(ptr))
-#define PTR_INT(ptr)  (long)(ptr)
+#define PTR_INT(ptr)  (uintptr_t)(ptr)
 
 #ifndef NULL
 #define NULL 0
